[Refactor] (판매자) 등록한 스토어 조회 API 리팩토링 (#621)
* [refactor] seller/sellers에 존재하던 판매자 스토어 조회 API를 seller/stores로 이동
- Seller가 Store를 등록하지 않은 상태인 경우 404 에러를 반환하도록 변경

* [refactor] API 리팩토링에 따른 사용하지 않는 DTO 제거

* [test] API 리팩토링에 따른 Test 코드 수정
