+++ src/main/java/com/bbangle/bbangle/seller/seller/controller/SellerController.java
@@ -1,7 +1,6 @@
 package com.bbangle.bbangle.seller.seller.controller;
 
 import com.bbangle.bbangle.common.dto.CommonResult;
-import com.bbangle.bbangle.common.dto.SingleResult;
 import com.bbangle.bbangle.common.service.ResponseService;
 import com.bbangle.bbangle.config.security.SellerApiPath;
 import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.AccountVerificationRequest;
@@ -9,7 +8,6 @@ import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.SellerAcco
 import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.SellerDocumentsRegisterRequest;
 import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.SellerStoreNameUpdateRequest;
 import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.SellerUpdateRequest;
-import com.bbangle.bbangle.seller.seller.controller.dto.SellerResponse.RegisteredStoreDetail;
 import com.bbangle.bbangle.seller.seller.controller.swagger.SellerApi;
 import com.bbangle.bbangle.seller.seller.facade.SellerFacade;
 import com.bbangle.bbangle.seller.seller.service.AccountVerificationService;
@@ -20,7 +18,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -38,17 +35,6 @@ public class SellerController implements SellerApi {
     private final SellerService sellerService;
     private final SellerFacade sellerFacade;
     private final AccountVerificationService accountVerificationService;
-
-    // TODO : v3 - Seller가 등록한 Store 정보 조회 (Store/Seller/Controller로 이동할 것)
-    @Override
-    @GetMapping()
-    public SingleResult<RegisteredStoreDetail> getRegisteredStoreDetail(
-        @AuthenticationPrincipal Long sellerId
-    ) {
-        return responseService.getSingleResult(
-            sellerFacade.getRegisteredStoreDetail(sellerId)
-        );
-    }
 
     @PostMapping(value = "/documents", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
     public CommonResult registerDocuments(
@@ -99,5 +85,4 @@ public class SellerController implements SellerApi {
         AccountVerificationInfo info = accountVerificationService.verifyAccount(request.toCommand(sellerId));
         return responseService.getSingleResult(info);
     }
-
 }
