+++ src/main/java/com/bbangle/bbangle/seller/seller/controller/swagger/SellerApi.java
@@ -1,14 +1,12 @@
 package com.bbangle.bbangle.seller.seller.controller.swagger;
 
 import com.bbangle.bbangle.common.dto.CommonResult;
-import com.bbangle.bbangle.common.dto.SingleResult;
 import com.bbangle.bbangle.exception.GlobalControllerAdvice;
 import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.AccountVerificationRequest;
 import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.SellerAccountUpdateRequest;
 import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.SellerDocumentsRegisterRequest;
 import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.SellerStoreNameUpdateRequest;
 import com.bbangle.bbangle.seller.seller.controller.dto.SellerRequest.SellerUpdateRequest;
-import com.bbangle.bbangle.seller.seller.controller.dto.SellerResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
@@ -21,14 +19,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 @Tag(name = "Seller", description = "(판매자) 판매자 API")
 public interface SellerApi {
-
-    @Operation(
-        summary = "판매자 스토어 정보 조회",
-        description = "판매자가 등록한 스토어의 상세 정보를 조회합니다."
-    )
-    SingleResult<SellerResponse.RegisteredStoreDetail> getRegisteredStoreDetail(
-        @AuthenticationPrincipal Long sellerId
-    );
 
     @Operation(
         summary = "(판매자) 판매자 서류 등록",
